Move subcommands to their own modules
Just code motion here.
