+++ src/options/add.rs
@@ -1,0 +1,17 @@
+use crate::{FuzzProject, RunCommand};
+use anyhow::Result;
+use structopt::StructOpt;
+
+#[derive(Clone, Debug, StructOpt)]
+pub struct Add {
+    #[structopt(required = true)]
+    /// Name of the new fuzz target
+    pub target: String,
+}
+
+impl RunCommand for Add {
+    fn run_command(&mut self) -> Result<()> {
+        let project = FuzzProject::new()?;
+        project.add_target(self)
+    }
+}
